requirements.txt
Flask
Werkzeug
opencv-python
ultralytics
moviepy
faiss-cpu
numpy
sentence-transformers
